backend/requirements.txt
# Django and REST framework
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-oauth-toolkit==2.3.0

# Database
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Production
gunicorn==21.2.0

# Security
django-filter==23.3
djangorestframework-simplejwt==5.3.0

# Utilities
Pillow==10.0.1  # For image processing if needed
whitenoise==6.5.0  # Static file serving